requirements.txt
numpy==1.22.3
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
pandas==1.4.1
Pillow==9.0.1
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
